jert_database/requirements.txt
mysql-connector-python
getpass